refactor: utilize the same url for fetching chart data
